rayFor() had some failures. x and y values where in a wront bracket. :)

+++ src/Camera/OrthographicCamera.java
@@ -42,18 +42,18 @@ public class OrthographicCamera extends Camera{
     public Ray rayFor(int w, int h, int x, int y) {
         // In order to calculate an orthographic ray, we need this formula:
         //
-        //                        (w-1)                        (h-1)
-        //                        -----                        -----
-        //                          2                            2
-        // o = e + a * s * ( x - --------- ) * u + s * ( y - --------- ) * v
-        //                        (w-1)                        (h-1)
+        //                        (w-1)                     (h-1)
+        //                     x- -----               y-    -----
+        //                          2                         2
+        // o = e + a * s * ( -------------- ) * u + s * ( --------------- ) * v
+        //                      (w-1)                        (h-1)
 
         // a is the aspectratio given by width and height
         final double a = (double) w / (double) h;
 
         // In order to ensure better readability we declare two variables standing for the two big brackets in the formular
-        final double xBracket = (double) x - ( (( (double)w-1.0) / 2.0 ) / ( (double) w - 1.0 ) );
-        final double yBracket = (double) y - ( (( (double)h-1.0) / 2.0 ) / ( (double) h - 1.0 ) );
+        final double xBracket = ( ( x - ( (double)w-1.0) / 2.0 ) / ( (double) w - 1.0 ) );
+        final double yBracket = ( ( y - ( (double)h-1.0) / 2.0 ) / ( (double) h - 1.0 ) );
 
         // Now the orthographic vector is calculated.
         final Point3 orthographicPosition = (this.e.add(this.u.mul((a*this.s*(xBracket))).add(this.v.mul(this.s*yBracket))));
